refactor(ui): extract SettingsDrawer normalizer + defaults (partial)
First-pass reduction of SettingsDrawer.tsx (1880 -> 1771 lines).
Moves the INLINE_DEFAULT_LINK_SETTINGS / CABLE_TEST_SETTINGS /
VULNERABILITY_SETTINGS constants, the withIds helper, and the
NormalizedTestsSettings type + normalizeTestsSettingsForSave
payload normaliser into a new settingsDrawerNormalizer.ts.

SettingsDrawer.tsx still exceeds the 800-line red-flag budget at
1771 lines and needs further work to break up the monolithic
component body. This commit isolates the pure-data extraction
that doesn't touch any component state.

Pure relocation, no behavior change.
